requirements.txt
# Dependencias principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Procesamiento de PDFs
PyPDF2==3.0.1
pdfplumber==0.10.3
python-magic==0.4.27


# OpenAI API
openai==1.3.9
python-dotenv==1.0.0

# Seguridad
bcrypt==4.0.1
passlib[bcrypt]==1.7.4

# Utilidades del sistema
python-magic==0.4.27
filetype==1.2.0
python-jose[cryptography]==3.3.0
